+++ tcex/__init__.py
@@ -1,6 +1,6 @@
-import os
-import sys
-import inspect
+# import os
+# import sys
+# import inspect
 
 __author__ = 'ThreatConnect (support@threatconnect.com)'
 __version__ = '0.0.12'
@@ -17,10 +17,10 @@ __license__ = 'Apache License, Version 2'
 ##         sys.path.insert(1, '%s/%s' % (lib_path, dirs.pop()))
 
 try:
-    from tcex import TcEx
+    from .tcex import TcEx
 except ImportError as e:
     print('Error: {}'.format(e))
     print('Try app.py --lib')
 
-from tcex_local import TcExLocal
-from argparser import ArgParser
+from .tcex_local import TcExLocal
+from .argparser import ArgParser